add mockdata to db
